+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release Notes
 
-## v1.2.0 - tbd
+## v1.3.0 - 2026-04-28
 
-- `readWithHeader()` in combination with `includeMissingCells()` will make sure that
+- `XlsxReader::getHeaders()` accepts parameter `worksheet` as worksheet name (string) or index (int)
+
+## v1.2.0 - 2026-04-19
+
+- `XlsxReader::readWithHeader()` in combination with `includeMissingCells()` will make sure that
   a cell for each header exists in each row.
 
 ## v1.1.0 - 2026-04-16